Back to localhost, refactor some classes
