+++ automation/src/test/scala/org/broadinstitute/dsde/workbench/page/ProxyRedirectPage.scala
@@ -1,14 +1,9 @@
-package org.broadinstitute.dsde.workbench.leonardo.notebooks
+package org.broadinstitute.dsde.workbench.page
 
 import org.broadinstitute.dsde.workbench.auth.AuthToken
-import org.broadinstitute.dsde.workbench.page.{CookieAuthedPage, PageUtil}
 import org.broadinstitute.dsde.workbench.service.test.WebBrowserUtil
 import org.openqa.selenium.WebDriver
-import org.scalatestplus.selenium.Page
 
-/**
- * Created by rtitle on 1/4/18.
- */
 class ProxyRedirectPage(override val url: String)(implicit val webDriver: WebDriver,
                                                   implicit override val authToken: AuthToken)
     extends CookieAuthedPage[ProxyRedirectPage]
